pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polnet"
version = "1.1.0"
description = "Python package for generating synthetic datasets of the cellular context for Cryo-Electron Tomography."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Antonio Martinez-Sanchez", email = "an.martinez.s.sw@gmail.com"},
    {name = "Juan Diego Gallego Nicolas", email = "jdiego.gallego@um.es"}
]
keywords = ["cryo-electron tomography", "synthetic data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "numpy>=1.26.3",
    "scipy>=1.11.4,<2",
    "mrcfile>=1.4.3,<2",
    "vtk>=9.3.0,<10",
    "pandas>=2.1.4,<3",
    "scikit-image>=0.22.0,<1",
    "PyYAML>=6.0",
    "torch>=2.0",
    "nibabel>=5.0",
]

[project.optional-dependencies]
gui = [
    "jupyter>=1.0.0",
    "ipyfilechooser>=0.6.0",
    "ipywidgets>=8.0",
    "tqdm>=4.0",
    "wget>=3.2",
]
dev = [
    "black>=24.0",
    "pylint>=3.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "sphinx-autodoc-typehints>=2.0",
]
all = ["polnet[gui,dev,docs]"]

[project.scripts]
polnet = "polnet.cli:app"

[project.urls]
Repository = "https://github.com/anmartinezs/polnet"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 79
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.idea
  | docs
  | data
)/
'''

[tool.pylint.format]
max-line-length = 79

[tool.pylint."messages control"]
disable = []

[tool.pylint.main]
extension-pkg-allow-list = ["vtk"]

[tool.pylint.master]
ignore-paths = ["data", "docs"]

[tool.pylint.typecheck]
ignored-modules = ["vtk"]
ignored-classes = ["vtk"]